move kind interface into extension registry module
