+++ src/backoffice/settings/section.manifests.ts
@@ -1,4 +1,4 @@
-import { ManifestSectionSidebarAppMenuKind } from '../shared/components/section/section-sidebar-menu/section-sidebar-menu.element';
+import { ManifestSectionSidebarAppMenuKind } from '@umbraco-cms/extensions-registry';
 import type { ManifestSection } from '@umbraco-cms/models';
 
 const sectionAlias = 'Umb.Section.Settings';
